.github/gitbook_worker/tools/requirements.txt
pandas
numpy
matplotlib
requests
tqdm
pygltflib
black
pyyaml
pytest
pytest-cov
beautifulsoup4
emoji
pathvalidate
markdown
pypandoc
